fix : outlierValue를 int 자료형에서 double 자료형으로 변경

+++ src/main/java/live/ioteatime/frontservice/dto/OutlierDto.java
@@ -16,6 +16,6 @@ public class OutlierDto {
     String place;
     String type;
     @JsonProperty("outlier_value")
-    long outlierValue;
+    double outlierValue;
     int flag;
 }
